refactor(site): separate data and rendering from index.ts (closes #5)
Extract typed data arrays to src/data.ts and pure rendering functions
to src/render.ts. Slim src/index.ts to routing and utility functions only.
Zero behavioral change: same routes, same output, same security headers.
